Rename testdata directory to linter_tests
