+++ testdata/src/linter_tests/ato.go
@@ -1,7 +1,7 @@
-package p
+package linter_tests
 
 // #nosec // want "Disabling of gosec must have an annotation associated with it. Please visit https://docs.google.com/document/d/1qiBNHlctSby0RZeaPzb-afVxAdA9vlrrQgce00zjDww/edit#heading=h.b2vss780hqfi"
-func nosecShouldHaveAnnotation(){}
+func nosecShouldHaveAnnotation() {}
 
 //RA Summary: [linter] - [linter type code] - [Linter summary] // want "Annotation needs approval from an ISSO"
 //RA: <Why did the linter flag this line of code?>
@@ -12,4 +12,4 @@ func nosecShouldHaveAnnotation(){}
 //RA Validator: jneuner@mitre.org
 //RA Modified Severity:
 // #nosec G100
-func whatHappens(){}
+func whatHappens() {}
